Remove warnings on singed/unsigned integers by converting OctaveFinder->size to int

+++ multiscale/octavefinder.hpp
@@ -31,7 +31,7 @@ namespace Colloids
             const double & get_prefactor() const {return this->prefactor;}
             const double & get_iterative_radius(const size_t l) const {return this->iterative_radii[l];};
             const double get_iterative_radius(const double &larger, const double &smaller) const;
-            const size_t & get_size(const size_t l) const {return this->sizes[l];};
+            const int & get_size(const size_t l) const {return this->sizes[l];};
             inline const cv::Mat_<bool> & get_binary(const size_t l) const {return binary[l-1];};
 			inline const Image & get_layersG(const size_t l) const {return layersG[l];}
 			inline const Image & get_layers(const size_t l) const {return layers[l];}
@@ -64,7 +64,7 @@ namespace Colloids
             std::vector<cv::Mat_<bool> > binary;
             std::vector<double> iterative_radii;
             std::vector<cv::FilterEngine> iterative_gaussian_filters;
-            std::vector<size_t> sizes;
+            std::vector<int> sizes;
             std::list<std::vector<int> > centers_no_subpix;
             double preblur_radius, prefactor;
             static std::map<size_t, cv::Mat_<double> > kernels;
